Feat: split Sidebar
