+++ src/components/Sidebar/index.tsx
@@ -1,22 +1,7 @@
-import { useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-
-const StyledOptionList = styled.aside<{ isOpen: boolean }>`
-  position: absolute;
-  top: 0;
-  right: 0;
-  width: 80%;
-  height: 100%;
-  transition: all 1s ease;
-  transform: ${(props) => (props.isOpen ? 'translateX(0)' : 'translateX(100%)')};
-`;
-
-const StyledOptionItem = styled.button`
-  width: 100%;
-  height: 1.5rem;
-  padding: 10px;
-`;
+import OptionItem from './OptionItem';
+import OptionList from './OptionList';
 
 type Option = {
   name: string;
@@ -24,13 +9,11 @@ type Option = {
 };
 
 export type SidebarProps = {
-  isOpen: boolean;
   onSidebarOpen: (isOpen: boolean) => void;
+  isOpen: boolean;
 };
 
 const Sidebar = ({ onSidebarOpen, isOpen }: SidebarProps) => {
-  const sidebarRef = useRef(null);
-
   const optionList: Array<Option> = [
     {
       name: '헬렌',
@@ -46,28 +29,14 @@ const Sidebar = ({ onSidebarOpen, isOpen }: SidebarProps) => {
     },
   ];
 
-  const onClickOutside = (event: any) => {
-    if (sidebarRef.current && !(sidebarRef.current as any).contains(event.target)) {
-      // TODO: change any type
-      return onSidebarOpen(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener('click', onClickOutside, true);
-    return () => {
-      document.removeEventListener('click', onClickOutside, true);
-    };
-  });
-
   return (
-    <StyledOptionList ref={sidebarRef} isOpen={isOpen}>
+    <OptionList onSidebarOpen={onSidebarOpen} isOpen={isOpen}>
       {optionList.map((option) => (
         <Link key={`${option.name}-${option.link}`} to={option.link}>
-          <StyledOptionItem>{option.name}</StyledOptionItem>
+          <OptionItem>{option.name}</OptionItem>
         </Link>
       ))}
-    </StyledOptionList>
+    </OptionList>
   );
 };
